requirements.txt
aniso8601==9.0.1
asn1crypto==1.5.1
certifi==2022.5.18.1
click==8.1.3
Flask==2.1.2
Flask-RESTful==0.3.9
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
pandas==1.3.5
pg8000==1.29.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
scramp==1.4.1
six==1.16.0
SQLAlchemy==1.4.37
typing_extensions==4.2.0
Werkzeug==2.1.2
zipp==3.8.0
